+++ application/tests/_ci_phpunit_test/CIPHPUnitTest.php
@@ -11,8 +11,14 @@
 class CIPHPUnitTest
 {
 	private static $loader_class = 'CI_Loader';
+	private static $autoload_dirs;
 
-	public static function init()
+	/**
+	 * Initialize CIPHPUnitTest
+	 * 
+	 * @param array $autoload_dirs directories to search class file for autoloader
+	 */
+	public static function init(array $autoload_dirs = null)
 	{
 		// Fix CLI args
 		$_server_backup = $_SERVER;
@@ -21,6 +27,8 @@ class CIPHPUnitTest
 			'welcome'	// Dummy
 		];
 		$_SERVER['argc'] = 2;
+
+		self::$autoload_dirs = $autoload_dirs;
 
 		// Load autoloader for ci-phpunit-test
 		require __DIR__ . '/autoloader.php';
@@ -65,6 +73,11 @@ class CIPHPUnitTest
 
 		// Restore $_SERVER. We need this for NetBeans
 		$_SERVER = $_server_backup;
+	}
+
+	public static function getAutoloadDirs()
+	{
+		return self::$autoload_dirs;
 	}
 
 	protected static function replaceLoader()